+++ Sources/VComponents/Components/Inputs/Code Entry View/VCodeEntryViewState.swift
@@ -1,0 +1,16 @@
+//
+//  VCodeEntryViewState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 02.09.23.
+//
+
+import Foundation
+import VCore
+
+// MARK: - V Code Entry View Internal State
+@available(iOS 15.0, *)
+@available(macOS, unavailable)
+@available(tvOS, unavailable)
+@available(watchOS, unavailable)
+typealias VCodeEntryViewInternalState = GenericState_EnabledFocusedDisabled